First order ZV-CV for the logistic example
The file hmc.repeats.cv.R is essentially a replacement to hmc.repeats.R. The first part is the same as hmc.repeats.R but with additional information stored. The second part involves performing four variants of ZV-CV. The "ub" and "ev" are estimators (i) and (ii) from our comment to the Jacob et al JRSS B read paper. They are based on minimising an upper bound for the variance and minimising the empirical variance, respectively.